Split NetworkX Cypher procedure backend
